Move out classes and functions from test files
