+++ src/3-interfaces/kitamura/searchCondition.ts
@@ -1,0 +1,3 @@
+export default interface SearchCondition {
+  status?: string[];
+}
