+++ src/3-interfaces/kitamura/searchParam.ts
@@ -1,0 +1,15 @@
+export default interface SearchParam {
+  status: string[];
+  shopName: string;
+  categoryName: string;
+  ownerName: string;
+  sequenceNum: string;
+  exportLineId: string;
+  dateFrom: string;
+  dateTo: string;
+  amountFrom: string;
+  amountTo: string;
+  scope: boolean;
+  departmentId: string;
+  includeChildDepartment: boolean;
+}
